[TTAHUB-4004/TTAHUB-4479] Collab report tables

+++ src/migrations/20250916193149-create-collab-report-tables.js
@@ -43,7 +43,7 @@ module.exports = {
           allowNull: false,
         },
         duration: {
-          type: Sequelize.FLOAT,
+          type: Sequelize.DOUBLE,
           allowNull: false,
           validate: {
             min: 0,
